chore: 예약하기 메서드를 예약서비스로 옮김
