+++ be/airbnb/src/main/java/com/codesquad/airbnb/web/controller/ReservationController.java
@@ -3,7 +3,7 @@ package com.codesquad.airbnb.web.controller;
 import com.codesquad.airbnb.web.config.annotation.CertifiedUser;
 import com.codesquad.airbnb.web.dto.ReservationPreview;
 import com.codesquad.airbnb.web.dto.UserInput;
-import com.codesquad.airbnb.web.service.rooms.RoomService;
+import com.codesquad.airbnb.web.service.reservation.ReservationService;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
 
@@ -12,14 +12,16 @@ import org.springframework.web.bind.annotation.*;
 @RequestMapping("/rooms/{roomId}/reservations")
 public class ReservationController {
 
-    private final RoomService roomService;
+    private final ReservationService reservationService;
 
-    public ReservationController(RoomService roomService) {
-        this.roomService = roomService;
+    public ReservationController(ReservationService reservationService) {
+        this.reservationService = reservationService;
     }
 
     @PostMapping
-    public ReservationPreview roomDetail(@PathVariable int roomId, @RequestBody UserInput userInput, @CertifiedUser int userId) {
-        return roomService.makeReservation(roomId, userId, userInput);
+    public ReservationPreview roomDetail(@PathVariable int roomId,
+                                         @RequestBody UserInput userInput,
+                                         @CertifiedUser int userId) {
+        return reservationService.makeReservation(roomId, userId, userInput);
     }
 }
